# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CodeLab – LeetCode-style competitive coding platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /problems:
    get:
      operationId: listProblems
      tags: [problems]
      summary: List problems with filters
      parameters:
        - name: difficulty
          in: query
          schema:
            type: string
            enum: [easy, medium, hard]
        - name: tag
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [solved, unsolved, attempted]
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of problems
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemListResponse"

    post:
      operationId: createProblem
      tags: [problems]
      summary: Create a new problem (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProblemInput"
      responses:
        "201":
          description: Problem created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
  /problems/{id}:
    get:
      operationId: getProblem
      tags: [problems]
      summary: Get problem detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Problem detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateProblem
      tags: [problems]
      summary: Update problem (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProblemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"

    delete:
      operationId: deleteProblem
      tags: [problems]
      summary: Delete problem (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── SUBMISSIONS ───────────────────────────────────────────────────────────
  /submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List submissions for the current user
      parameters:
        - name: problemId
          in: query
          schema:
            type: integer
        - name: language
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Submissions list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionListResponse"

    post:
      operationId: submitCode
      tags: [submissions]
      summary: Submit code for a problem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "202":
          description: Submission accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}:
    get:
      operationId: getSubmission
      tags: [submissions]
      summary: Get submission result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/run:
    post:
      operationId: runCode
      tags: [submissions]
      summary: Run code against custom test input (no submission)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunCodeInput"
      responses:
        "200":
          description: Run result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunResult"

  # ── CONTESTS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    TestCase:
      type: object
      required: [input, expectedOutput]
      properties:
        input:          { type: string }
        expectedOutput: { type: string }
        isHidden:       { type: boolean }
    Problem:
      type: object
      required: [id, title, slug, difficulty, description, tags, createdAt, solvedCount, acceptanceRate]
      properties:
        id:              { type: integer }
        title:           { type: string }
        slug:            { type: string }
        difficulty:      { type: string, enum: [easy, medium, hard] }
        description:     { type: string }
        constraints:     { type: ["string", "null"] }
        hints:           { type: ["string", "null"] }
        tags:            { type: array, items: { type: string } }
        starterCode:     { type: ["string", "null"] }
        testCases:       { type: array, items: { $ref: "#/components/schemas/TestCase" } }
        createdAt:       { type: string }
        solvedCount:     { type: integer }
        acceptanceRate:  { type: number }
        userStatus:      { type: ["string", "null"] }
    ProblemInput:
      type: object
      required: [title, difficulty, description, tags]
      properties:
        title:       { type: string }
        difficulty:  { type: string, enum: [easy, medium, hard] }
        description: { type: string }
        constraints: { type: string }
        hints:       { type: string }
        tags:        { type: array, items: { type: string } }
        starterCode: { type: string }
        testCases:   { type: array, items: { $ref: "#/components/schemas/TestCase" } }
    ProblemUpdate:
      type: object
      properties:
        title:       { type: string }
        difficulty:  { type: string, enum: [easy, medium, hard] }
        description: { type: string }
        constraints: { type: string }
        hints:       { type: string }
        tags:        { type: array, items: { type: string } }
        starterCode: { type: string }
        testCases:   { type: array, items: { $ref: "#/components/schemas/TestCase" } }
    ProblemListResponse:
      type: object
      required: [problems, total, page, limit]
      properties:
        problems: { type: array, items: { $ref: "#/components/schemas/Problem" } }
        total:    { type: integer }
        page:     { type: integer }
        limit:    { type: integer }

    # SUBMISSIONS
    Submission:
      type: object
      required: [id, problemId, userId, language, code, status, createdAt]
      properties:
        id:           { type: integer }
        problemId:    { type: integer }
        userId:       { type: integer }
        language:     { type: string }
        code:         { type: string }
        status:       { type: string, enum: [pending, running, accepted, wrong_answer, time_limit_exceeded, runtime_error, compilation_error] }
        runtime:      { type: ["integer", "null"] }
        memory:       { type: ["integer", "null"] }
        output:       { type: ["string", "null"] }
        errorMessage: { type: ["string", "null"] }
        createdAt:    { type: string }
        problemTitle: { type: ["string", "null"] }
    SubmissionInput:
      type: object
      required: [problemId, language, code]
      properties:
        problemId:  { type: integer }
        language:   { type: string }
        code:       { type: string }
        contestId:  { type: ["integer", "null"] }
    SubmissionListResponse:
      type: object
      required: [submissions, total, page, limit]
      properties:
        submissions: { type: array, items: { $ref: "#/components/schemas/Submission" } }
        total:       { type: integer }
        page:        { type: integer }
        limit:       { type: integer }
    RunCodeInput:
      type: object
      required: [language, code, input]
      properties:
        language: { type: string }
        code:     { type: string }
        input:    { type: string }
    RunResult:
      type: object
      required: [status, output]
      properties:
        status:      { type: string }
        output:      { type: string }
        stderr:      { type: ["string", "null"] }
        runtime:     { type: ["integer", "null"] }
        memory:      { type: ["integer", "null"] }

    # CONTESTS
...