# lib/api-zod/src/generated/types/problem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeLab – LeetCode-style competitive coding platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProblemDifficulty } from './problemDifficulty';
import type { TestCase } from './testCase';

export interface Problem {
  id: number;
  title: string;
  slug: string;
  difficulty: ProblemDifficulty;
  description: string;
  /** @nullable */
  constraints?: string | null;
  /** @nullable */
  hints?: string | null;
  tags: string[];
  /** @nullable */
  starterCode?: string | null;
  testCases?: TestCase[];
  createdAt: string;
  solvedCount: number;
  acceptanceRate: number;
  /** @nullable */
  userStatus?: string | null;
}