# lib/api-zod/src/generated/types/problemListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeLab – LeetCode-style competitive coding platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Problem } from './problem';

export interface ProblemListResponse {
  problems: Problem[];
  total: number;
  page: number;
  limit: number;
}