# lib/api-zod/src/generated/types/problemUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeLab – LeetCode-style competitive coding platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProblemUpdateDifficulty } from './problemUpdateDifficulty';
import type { TestCase } from './testCase';

export interface ProblemUpdate {
  title?: string;
  difficulty?: ProblemUpdateDifficulty;
  description?: string;
  constraints?: string;
  hints?: string;
  tags?: string[];
  starterCode?: string;
  testCases?: TestCase[];
}