# lib/api-zod/src/generated/types/submissionListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CodeLab – LeetCode-style competitive coding platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Submission } from './submission';

export interface SubmissionListResponse {
  submissions: Submission[];
  total: number;
  page: number;
  limit: number;
}